move localstorage.ts to excalidraw-app
